removed checkout component
