+++ src/components/Cart.js
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from 'react';
-import { NavLink } from 'react-router-dom';
 import { getCartByUser, deleteOrderProduct, cancelledOrder, completedOrder } from '../api/index.js'
 
 const Cart = (props) => {
@@ -7,31 +6,31 @@ const Cart = (props) => {
     const {shoppingCart, setShoppingCart} = props
     const {user, token, setOrderId } = props
 
-    // const handleCompleteOrder = async (id) => {
-    //     try {
-    //         // event.preventDefault();
-    //         const result = await completedOrder(id, token) 
-    //         console.log('resultcancellation', result)
-    //         // console.log("What is RESULT ERROR", result.error)
-    //         // console.log("What is ERROR MESSAGE", result.data.message)
-    //         // console.log("What is RESULT ERROR", { error: result.data.error})
-    //     } catch(error) {
-    //         console.error(error)
-    //     }
-    // };
+    const handleCompleteOrder = async (id) => {
+        try {
+            // event.preventDefault();
+            const result = await completedOrder(id, token) 
+            console.log('resultcancellation', result)
+            // console.log("What is RESULT ERROR", result.error)
+            // console.log("What is ERROR MESSAGE", result.data.message)
+            // console.log("What is RESULT ERROR", { error: result.data.error})
+        } catch(error) {
+            console.error(error)
+        }
+    };
 
-    // const handleCancelOrder = async (id) => {
-    //     try {
-    //         // event.preventDefault();
-    //         const result = await cancelledOrder(id, token) 
-    //         console.log('resultcancellation', result)
-    //         // console.log("What is RESULT ERROR", result.error)
-    //         // console.log("What is ERROR MESSAGE", result.data.message)
-    //         // console.log("What is RESULT ERROR", { error: result.data.error})
-    //     } catch(error) {
-    //         console.error(error)
-    //     }
-    // };
+    const handleCancelOrder = async (id) => {
+        try {
+            // event.preventDefault();
+            const result = await cancelledOrder(id, token) 
+            console.log('resultcancellation', result)
+            // console.log("What is RESULT ERROR", result.error)
+            // console.log("What is ERROR MESSAGE", result.data.message)
+            // console.log("What is RESULT ERROR", { error: result.data.error})
+        } catch(error) {
+            console.error(error)
+        }
+    };
 
     useEffect(() => {
         getCartByUser(token)
@@ -92,7 +91,8 @@ const Cart = (props) => {
                                 <button id={product.orderProductId} type="submit" onClick={handleRemove}>Remove From Cart</button>
                             </div>)
                         }
-                            <NavLink to="/cart/checkout">Continue to Checkout</NavLink>
+                        <button type="submit" onClick={() => handleCancelOrder(shoppingCart.id)}>cancel Order</button>
+                        <button type="submit" onClick={() => handleCompleteOrder(shoppingCart.id)}>Complete Order</button>
                         </section>
                         </>
                         : ''
